Update transformers_models.py
change default dtype from float16 to float32

+++ optimum/quanto/models/transformers_models.py
@@ -138,7 +138,7 @@ class QuantizedTransformersModel(ModelHubMixin):
         config = AutoConfig.from_pretrained(pretrained_model_name_or_path, **kwargs)
         with init_empty_weights():
             model = cls.auto_class.from_config(config)
-            dtype = kwargs.get("torch_dtype", torch.float16)
+            dtype = kwargs.get("torch_dtype", torch.float32)
             model = model.to(dtype=dtype)
         # Look for the index of a sharded checkpoint
         checkpoint_file = os.path.join(working_dir, SAFE_WEIGHTS_INDEX_NAME)
